+++ CHANGELOG.md
@@ -1,14 +1,17 @@
 # CHANGELOG
 
 ## [Planning]
-- allow xslx format (excel file).[next version]
+- allow xslx format (excel file).[miner version]
 - deploy heroku automaticaly by `heroku button`.[patch]
-- dynamic channing flat file class in settings[patch]
-
+- dynamic setting flat file class in settings[patch]
+<!-- If about current planning all done then it next major version
+ -->
+ 
 ## 0.3.0
 - uploading flat file with celery(paytm only).
 - Update code proper message in new entry.
 - Adding package.serializers.ItemsListSerializer default to date fields manully.
+- Optional cache system in air pollution system.
 
 ## 0.2.1
 - Changing the `packages.view` into readable sub folders.